ICMSLST-671 Make WithdrawApplication re-usable between import/export.
