+++ web/domains/case/_import/models.py
@@ -298,41 +298,6 @@ class ImportContact(models.Model):
     updated_datetime = models.DateTimeField(auto_now=True)
 
 
-class WithdrawImportApplication(models.Model):
-    STATUS_OPEN = "open"
-    STATUS_REJECTED = "rejected"
-    STATUS_ACCEPTED = "accepted"
-
-    OPEN = (STATUS_OPEN, "OPEN")
-    REJECTED = (STATUS_REJECTED, "REJECTED")
-    ACCEPTED = (STATUS_ACCEPTED, "ACCEPTED")
-    STATUSES = (OPEN, REJECTED, ACCEPTED)
-
-    import_application = models.ForeignKey(
-        ImportApplication, on_delete=models.PROTECT, related_name="withdrawals"
-    )
-    is_active = models.BooleanField(default=True)
-    status = models.CharField(max_length=10, choices=STATUSES, default=STATUS_OPEN)
-    reason = models.TextField()
-    request_by = models.ForeignKey(
-        User,
-        on_delete=models.PROTECT,
-        related_name="+",
-    )
-
-    response = models.TextField()
-    response_by = models.ForeignKey(
-        User,
-        on_delete=models.PROTECT,
-        blank=True,
-        null=True,
-        related_name="+",
-    )
-
-    created_datetime = models.DateTimeField(auto_now_add=True)
-    updated_datetime = models.DateTimeField(auto_now=True)
-
-
 class EndorsementImportApplication(models.Model):
     import_application = models.ForeignKey(
         ImportApplication, on_delete=models.PROTECT, related_name="endorsements"
